requirements.txt
cohere==5.15.0
fastapi==0.115.12
httpx==0.28.1
numpy==2.2.5
pandas==2.2.3
pydantic==2.11.4
pytest==8.3.5
python-dotenv==1.1.0
uvicorn==0.34.2
weaviate_client==4.12.0
